Rename (library) : 폴더 참조 경로 업데이트
